+++ inference-engine/tests/ngraph_functions/include/ngraph_functions/builders.hpp
@@ -385,8 +385,8 @@ std::shared_ptr<ngraph::Node> makeConcat(const std::vector<ngraph::Output<Node>>
                                          const int& axis);
 
 std::shared_ptr<ngraph::Node> makePad(const ngraph::Output<Node>& data,
-                                      const std::vector<size_t>& padsBegin,
-                                      const std::vector<size_t>& padsEnd,
+                                      const std::vector<int64_t>& padsBegin,
+                                      const std::vector<int64_t>& padsEnd,
                                       float argPadValue,
                                       ngraph::helpers::PadMode padMode);
